+++ app/src/main/java/eu/qm/fiszki/model/flashcard/Flashcard.kt
@@ -1,6 +1,8 @@
 package eu.qm.fiszki.model.flashcard
 
 import com.j256.ormlite.field.DatabaseField
+import eu.qm.fiszki.algorithm.fsrs.FsrsCard
+import eu.qm.fiszki.algorithm.fsrs.FsrsState
 import java.io.Serializable
 
 class Flashcard : Serializable {
@@ -35,6 +37,33 @@ class Flashcard : Serializable {
     @DatabaseField
     var staticPass: Int = 0
         private set
+
+    @DatabaseField(columnName = "fsrsStability", defaultValue = "0.0")
+    var fsrsStability: Double = 0.0
+
+    @DatabaseField(columnName = "fsrsDifficulty", defaultValue = "0.0")
+    var fsrsDifficulty: Double = 0.0
+
+    @DatabaseField(columnName = "fsrsElapsedDays", defaultValue = "0")
+    var fsrsElapsedDays: Int = 0
+
+    @DatabaseField(columnName = "fsrsScheduledDays", defaultValue = "0")
+    var fsrsScheduledDays: Int = 0
+
+    @DatabaseField(columnName = "fsrsReps", defaultValue = "0")
+    var fsrsReps: Int = 0
+
+    @DatabaseField(columnName = "fsrsLapses", defaultValue = "0")
+    var fsrsLapses: Int = 0
+
+    @DatabaseField(columnName = "fsrsState", defaultValue = "0")
+    var fsrsState: Int = 0
+
+    @DatabaseField(columnName = "fsrsLastReview", defaultValue = "0")
+    var fsrsLastReview: Long = 0L
+
+    @DatabaseField(columnName = "fsrsLastRating", defaultValue = "0")
+    var fsrsLastRating: Int = 0
 
     constructor()
 
@@ -83,5 +112,27 @@ class Flashcard : Serializable {
     fun resetStatictic() {
         staticFail = 0
         staticPass = 0
+    }
+
+    fun toFsrsCard(): FsrsCard = FsrsCard(
+        stability = fsrsStability,
+        difficulty = fsrsDifficulty,
+        elapsedDays = fsrsElapsedDays,
+        scheduledDays = fsrsScheduledDays,
+        reps = fsrsReps,
+        lapses = fsrsLapses,
+        state = FsrsState.entries[fsrsState],
+        lastReview = fsrsLastReview
+    )
+
+    fun applyFsrsCard(card: FsrsCard) {
+        fsrsStability = card.stability
+        fsrsDifficulty = card.difficulty
+        fsrsElapsedDays = card.elapsedDays
+        fsrsScheduledDays = card.scheduledDays
+        fsrsReps = card.reps
+        fsrsLapses = card.lapses
+        fsrsState = card.state.ordinal
+        fsrsLastReview = card.lastReview
     }
 }